pyproject.toml
[tool.poetry]
name = "printacc"
version = "0.1.0"
description = "Print Accounting System"
authors = ["Alex Rozum <rozumalex@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.6.*"
flake8 = "^3.8.4"
pytest = "^6.1.1"
Django = "2.2"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.1"
Markdown = "^3.3"
django-filter = "^2.4.0"
django-pytest = "^0.2.0"
pytest-cov = "^2.10.1"
django-rest-swagger = "^2.2.0"
codecov = "^2.1.10"
django-rest-auth = "^0.9.5"
django-rest-authtoken = "^2.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
